+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * It is possible to navigate the menu using tab
   * Fixed bug where stderr was not being displayed
   * Vue components are now described using single file components syntax
+  * Added Vuex store to handle app state
 
 0.4.0 / 2017-10-02
 ==================